+++ doc/changelog.md
@@ -1,5 +1,6 @@
 # v2.2.1-dev
-_Changes: faster performance for logged in users, upgrade Flask to 0.12_
+_Changes: faster performance for logged in users, upgrade Flask to 0.12,
+sort files and folders case-insensitively in web interface_
 
 A new caching layer for user-session data has reduced the typical per-request
 time for small images from 9ms to 3ms, which is now in line with the performance